refactor: 공통 Header 용도별 컴포넌트 분리
