+++ src/components/header/CompactHeader.tsx
@@ -1,0 +1,7 @@
+import type { FC } from "react";
+
+const CompactHeader: FC = () => {
+  return <div></div>;
+};
+
+export default CompactHeader;
