+++ src/layouts/AuthLayout.tsx
@@ -1,5 +1,5 @@
 import type { FC, ReactNode } from "react";
-import Header from "@/components/common/Header.tsx";
+import Header from "@/components/header/Header.tsx";
 
 type LayoutProps = {
   children: ReactNode;
@@ -8,7 +8,7 @@ type LayoutProps = {
 const AuthLayout: FC<LayoutProps> = ({ children }) => {
   return (
     <div className="flex flex-col w-full min-w-0 flex-1">
-      <Header variant="auth"></Header>
+      <Header variant="auth" />
       <main>{children}</main>
     </div>
   );
